feat: 댓글 기능 예외처리
- GlobalExceptionHandler에 @Valid가 발생한 예외 처리
- 댓글 기능에 CustomException throw

+++ src/main/java/com/theocean/fundering/domain/comment/service/CommentService.java
@@ -5,6 +5,10 @@ import com.theocean.fundering.domain.member.domain.Member;
 import com.theocean.fundering.domain.comment.dto.CommentRequest;
 import com.theocean.fundering.domain.comment.dto.CommentResponse;
 import com.theocean.fundering.domain.comment.repository.CommentRepository;
+import com.theocean.fundering.global.errors.exception.Exception400;
+import com.theocean.fundering.global.errors.exception.Exception403;
+import com.theocean.fundering.global.errors.exception.Exception404;
+import com.theocean.fundering.global.errors.exception.Exception500;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -26,15 +30,15 @@ public class CommentService {
 
     // (기능) 댓글 작성
     @Transactional
-    public Comment createComment(Long memberId, Long postId, CommentRequest.saveDTO request) {
+    public void createComment(Long memberId, Long postId, CommentRequest.saveDTO request) {
 
         // 댓글 작성 이전 회원과 게시물 존재여부 확인
         if (!memberRepository.existsById(memberId)) {
-            throw new IllegalArgumentException("존재하지 않는 회원입니다: " + memberId);
+            throw new Exception400("존재하지 않는 회원입니다: " + memberId);
         }
 
         if (!postRepository.existsById(postId)) {
-            throw new IllegalArgumentException("존재하지 않는 게시물입니다: " + postId);
+            throw new Exception404("해당 게시글을 찾을 수 없습니다: " + postId);
         }
 
         // 댓글 순서 구하는 로직 - 특정 post내에서 가장 최근에 작성된 댓글의 순서 + 1
@@ -55,7 +59,7 @@ public class CommentService {
         // 대댓글 생성 로직 (부모 댓글이 존재할 경우)
         if (parentCommentOrder != null) {
             Comment parentComment = commentRepository.findByPostIdAndCommentOrder(postId, parentCommentOrder)
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다: " + parentCommentOrder));
+                    .orElseThrow(() -> new Exception400("존재하지 않는 댓글입니다: " + parentCommentOrder));
 
             newComment.updateParentCommentOrder(parentCommentOrder);
 
@@ -69,13 +73,23 @@ public class CommentService {
             newComment.updateParentCommentOrder(commentOrder);
         }
 
-        return commentRepository.save(newComment);
+        // postId와 commentOrder를 기반으로 중복 검사
+        if (commentRepository.existsByPostIdAndCommentOrder(postId, commentOrder)) {
+            throw new Exception400("댓글 순번이 중복되었습니다: " + commentOrder);
+        }
+
+        try {
+            commentRepository.save(newComment);
+        } catch(Exception e) {
+            throw new Exception500("댓글 저장 중 문제가 발생했습니다.");
+        }
+
     }
 
     // (기능) 댓글 목록 조회 - 유저Id를 이용하여 commentsDTO를 생성하는 로직
     private CommentResponse.commentsDTO createCommentsDTO(Comment comment) {
         Member writer = memberRepository.findById(comment.getWriterId())
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다: " + comment.getWriterId()));
+                .orElseThrow(() -> new Exception404("존재하지 않는 회원입니다: " + comment.getWriterId()));
 
         return new CommentResponse.commentsDTO(
                 comment,
@@ -88,7 +102,19 @@ public class CommentService {
 
     // (기능) 댓글 목록 조회 - 컨트롤러로 findAllDTO 리턴
     public CommentResponse.findAllDTO getCommentsDtoByPostId(long postId, PageRequest pageRequest) {
-        Page<Comment> commentsPage = commentRepository.findByPostIdOrdered(postId, pageRequest);
+
+        // 게시글 존재 여부 확인
+        if (!postRepository.existsById(postId)) {
+            throw new Exception404("해당 게시글을 찾을 수 없습니다: " + postId);
+        }
+
+        // 댓글 페이징
+        Page<Comment> commentsPage;
+        try {
+            commentsPage = commentRepository.findByPostIdOrdered(postId, pageRequest);
+        }catch(Exception e) {
+            throw new Exception500("댓글 조회 도중 문제가 발생했습니다.");
+        }
 
 
         List<CommentResponse.commentsDTO> commentsDtos = commentsPage.getContent().stream()
@@ -101,10 +127,24 @@ public class CommentService {
 
     // (기능) 댓글 삭제
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long memberId, Long postId, Long commentId) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다: " + commentId));
+                .orElseThrow(() -> new Exception404("존재하지 않는 댓글입니다: " + commentId));
 
-        commentRepository.delete(comment);
+        // 게시글 존재 여부 확인
+        if (!postRepository.existsById(postId)) {
+            throw new Exception404("해당 게시글을 찾을 수 없습니다: " + postId);
+        }
+
+        // 권한 확인
+        if (!comment.getWriterId().equals(memberId)) {
+            throw new Exception403("댓글 삭제 권한이 없습니다.");
+        }
+
+        try {
+            commentRepository.delete(comment);
+        } catch (Exception e) {
+            throw new Exception500("댓글 삭제처리 도중 문제가 발생했습니다.");
+        }
     }
 }
